+++ src/post-processing/imagemagick-post-processor.vala
@@ -24,20 +24,27 @@ namespace Peek.PostProcessing {
 
         double delay = (100.0 / framerate);
         var output_file = Utils.create_temp_file ("gif");
+        var temp_dir = Utils.get_temp_dir ();
+
+        string? magick_debug = Environment.get_variable ("PEEK_MAGICK_DEBUG");
+        if (Utils.string_is_empty (magick_debug)) {
+          magick_debug = "None";
+        }
+
+        debug ("Running ImageMagick convert, saving to %s, using temporary path %s",
+          output_file, temp_dir);
+
         string[] argv = {
           "convert",
-          // "-debug", "All",
+          "-debug", magick_debug,
           "-set", "delay", delay.to_string (),
           "-layers", "Optimize",
+          "-define", "registry:temporary-path=" + temp_dir,
           file.get_path (),
           output_file
         };
 
-        debug ("Running ImageMagick convert, saving to %s", output_file);
-        string[] env = Environ.get ();
-        Environ.set_variable (env, "TMPDIR", Utils.get_temp_dir (), true);
-        Environ.unset_variable (env, "MAGICK_TMPDIR");
-        Process.spawn_async (null, argv, env,
+        Process.spawn_async (null, argv, null,
           SpawnFlags.SEARCH_PATH | SpawnFlags.DO_NOT_REAP_CHILD, null, out pid);
 
         ChildWatch.add (pid, (pid, status) => {