+++ CHANGES.md
@@ -2,6 +2,7 @@
  * misc: Use GtkFileChooserNative, if available. Most importantly this allows
          saving the files everywhere when Peek is running as Flatpak package
          without giving Peek access to the entire file system.
+ * fix: Set temporary directory for ImageMagick
 
 # Version 1.0.0 - 2017-03-11
  * recording: Support GNOME Shell screencast DBUS service. Allows recording